Player Coins Rounding Fix
- Fixed the method which animates the UI Coin counter where exact values between integers and floats are compared. It has also been optimized.

+++ LineGame/Assets/Scripts/Levels/LevelManager.cs
@@ -350,7 +350,7 @@ public class LevelManager : Singleton<LevelManager>
 
 		// Count up the UI element from the previous Coin count to new coin count over time
 		// Cache the now-old coin count
-		float oldCointCount = GameSave.CoinCount;   
+		int oldCointCount = GameSave.CoinCount;   
 		ApplyLevelReward();
 
 		// Update the Coin Count Display
